chore(k-table): merge kt-table and kt-standard-table directories
