+++ packages/kotti-ui/source/index.ts
@@ -91,9 +91,7 @@ import { KtPopover } from './kotti-popover'
 export * from './kotti-popover'
 import { KtRow } from './kotti-row'
 export * from './kotti-row'
-import { KtStandardTable } from './kotti-standard-table'
-export * from './kotti-standard-table'
-import { KtTable } from './kotti-table'
+import { KtStandardTable, KtTable } from './kotti-table'
 export * from './kotti-table'
 import {
 	KtTableLegacy,
