+++ packages/kotti-ui/source/kotti-i18n/types.ts
@@ -10,7 +10,7 @@ import type { Shared as KottiFieldSelectShared } from '../kotti-field-select/typ
 import type { KottiFilters } from '../kotti-filters/types'
 import type { KottiFormSubmit } from '../kotti-form-submit/types'
 import type { KottiNavbar } from '../kotti-navbar/types'
-import type { KottiStandardTable } from '../kotti-standard-table/types'
+import type { KottiStandardTable } from '../kotti-table/standard-table/types'
 import type { KottiValueLabel } from '../kotti-value-label/types'
 import { DecimalSeparator } from '../types/decimal-separator'
