+++ packages/kotti-ui/source/kotti-table/index.ts
@@ -2,12 +2,15 @@ import type { Kotti } from '../types'
 import { MetaDesignType } from '../types/kotti'
 import { attachMeta, makeInstallable } from '../utilities'
 
-import { useKottiTable as _useKottiTable } from './hooks'
+import KtStandardTableVue from './KtStandardTable.vue'
 import KtTableVue from './KtTable.vue'
-import { KottiTable } from './types'
+import { useKottiStandardTable as _useKottiStandardTable } from './standard-table/hooks'
+import { KottiStandardTable } from './standard-table/types'
+import { useKottiTable as _useKottiTable } from './table/hooks'
+import { KottiTable } from './table/types'
 
-const META: Kotti.Meta = {
-	addedVersion: '1337.0.0.1',
+const TABLE_META: Kotti.Meta = {
+	addedVersion: '7.4.0',
 	deprecated: null,
 	designs: {
 		type: MetaDesignType.FIGMA,
@@ -47,5 +50,32 @@ const META: Kotti.Meta = {
 	},
 }
 
-export const KtTable = attachMeta(makeInstallable(KtTableVue), META)
+const STANDARD_META: Kotti.Meta = {
+	addedVersion: '7.4.0',
+	deprecated: null,
+	designs: {
+		type: MetaDesignType.FIGMA,
+		url: 'https://www.figma.com/design/0yFVivSWXgFf2ddEF92zkf/Kotti-Design-System?node-id=6305-10646&node-type=canvas&t=8lzEM5nlkrh8aUMF-0',
+	},
+	slots: {
+		buttons: { description: 'slot next to table actions', scope: null },
+		'header-action': { description: 'slot next to title', scope: null },
+		'switch-buttons': {
+			description: 'slot next to applied filters',
+			scope: null,
+		},
+	},
+	typeScript: {
+		namespace: 'Kotti.StandardTable',
+		schema: KottiStandardTable.propsSchema,
+	},
+}
+
+export const KtTable = attachMeta(makeInstallable(KtTableVue), TABLE_META)
 export const useKottiTable = _useKottiTable
+
+export const KtStandardTable = attachMeta(
+	makeInstallable(KtStandardTableVue),
+	STANDARD_META,
+)
+export const useKottiStandardTable = _useKottiStandardTable
