+++ packages/kotti-ui/source/kotti-table/standard-table/components/Filters.vue
@@ -35,8 +35,8 @@
 import { defineComponent } from 'vue'
 import { Yoco } from '@3yourmind/yoco'
 
-import { useTranslationNamespace } from '../../kotti-i18n/hooks'
-import { makeProps } from '../../make-props'
+import { useTranslationNamespace } from '../../../kotti-i18n/hooks'
+import { makeProps } from '../../../make-props'
 
 import { KottiStandardTable } from '../types'
 import { getDefaultValue, isFilterValueEmpty } from '../utilities/filters'
