+++ packages/kotti-ui/source/kotti-table/standard-table/components/Search.vue
@@ -17,10 +17,10 @@ import debounce from 'lodash/debounce'
 import { defineComponent, ref, watch } from 'vue'
 import { Yoco } from '@3yourmind/yoco'
 
-import { DEFAULT_DEBOUNCE } from '../../constants'
-import { useTranslationNamespace } from '../../kotti-i18n/hooks'
-import type { KottiFieldText } from '../../kotti-field-text/types'
-import { makeProps } from '../../make-props'
+import { DEFAULT_DEBOUNCE } from '../../../constants'
+import { useTranslationNamespace } from '../../../kotti-i18n/hooks'
+import type { KottiFieldText } from '../../../kotti-field-text/types'
+import { makeProps } from '../../../make-props'
 
 import { KottiStandardTable } from '../types'
