+++ packages/kotti-ui/source/kotti-table/standard-table/components/filters/NumberRange.vue
@@ -24,9 +24,9 @@
 import { computed, defineComponent, ref, watch } from 'vue'
 import isEqual from 'lodash/isEqual'
 
-import { useTranslationNamespace } from '../../../kotti-i18n/hooks'
-import type { KottiFieldNumber } from '../../../kotti-field-number/types'
-import { makeProps } from '../../../make-props'
+import { useTranslationNamespace } from '../../../../kotti-i18n/hooks'
+import type { KottiFieldNumber } from '../../../../kotti-field-number/types'
+import { makeProps } from '../../../../make-props'
 
 import { KottiStandardTable } from '../../types'
 import { reOrderRange } from '../../utilities/filters'
