+++ packages/kotti-ui/source/kotti-table/standard-table/utilities/filters.ts
@@ -1,13 +1,13 @@
 import { Dashes } from '@metatypes/typography'
 
-import { useTranslationNamespace } from '../../kotti-i18n/hooks'
-import type { KottiFieldDateRange } from '../../kotti-field-date/types'
-import type { KottiFieldNumber } from '../../kotti-field-number/types'
+import { useTranslationNamespace } from '../../../kotti-i18n/hooks'
+import type { KottiFieldDateRange } from '../../../kotti-field-date/types'
+import type { KottiFieldNumber } from '../../../kotti-field-number/types'
 import type {
 	KottiFieldMultiSelect,
 	KottiFieldSingleSelect,
-} from '../../kotti-field-select/types'
-import type { KottiFieldToggle } from '../../kotti-field-toggle/types'
+} from '../../../kotti-field-select/types'
+import type { KottiFieldToggle } from '../../../kotti-field-toggle/types'
 
 import { KottiStandardTable } from '../types'
