+++ packages/kotti-ui/source/kotti-table/standard-table/utilities/translation.ts
