+++ packages/kotti-ui/source/types/kotti.ts
@@ -66,8 +66,8 @@ export { KottiPagination as Pagination } from '../kotti-pagination/types'
 export { KottiPopover as Popover } from '../kotti-popover/types'
 export { KottiRow as Row } from '../kotti-row/types'
 export { KottiTableLegacy as TableLegacy } from '../kotti-table-legacy/types'
-export { KottiTable as Table } from '../kotti-table/types'
-export { KottiStandardTable as StandardTable } from '../kotti-standard-table/types'
+export { KottiTable as Table } from '../kotti-table/table/types'
+export { KottiStandardTable as StandardTable } from '../kotti-table/standard-table/types'
 export { KottiTag as Tag } from '../kotti-tag/types'
 export { KottiUserMenu as UserMenu } from '../kotti-user-menu/types'
 export { KottiValueLabel as ValueLabel } from '../kotti-value-label/types'
